Implemented Sprinting and Stamina.
Implemented a sprint mechanic so that if the player presses shift while moving on the ground the player sprints and moves faster by a multiplier.  The player should be above the staminaThresholdStartSprinting to be able to start sprinting. While sprinting stamina drops by the staminaDegenRateSprinting and once it hits a certain threshold the player can no longer sprint until a staminaRegenDelay passes, and sprinting regenerates by the staminaRegenRate past the staminaThresholdStartSprinting.